extract common logic to a module
